# public/openapi.yaml
info:
  description: |
    Purpie is an open source social media solution focusing on video oriented communication and content sharing.
    
    The Purpie endpoints are offered to our users to facilitate the use of our technology and allow seamless integrations with their applications.
    For this documentation, you can use the **Api Key** and **Api Secret** which generated from [Purpie](https://purpie.io/) settings page to test the services.

    # Introduction
    Purpie orchestrates various media services and solutions for online meetings, live streamings, on demand video delivery and many more.
    With an agile and hierarchical way of user interaction and content sharing possibilities, Purpie proposes a robust and flexible solution for
    various video communication and distribution needs.

    # Cross-Origin Resource Sharing
    This API features Cross-Origin Resource Sharing (CORS) implemented.
    And that allows cross-domain communication from the browser.
    All responses have a wildcard same-origin which makes them completely public and accessible to everyone, including any code on any site.

    # Authentication
    To make authorized calls on Purpie API operations, you need to generate an API key and secret in the settings page after login to [Purpie](https://purpie.io/).
    
    After generating API credentials, you need to note your `api secret`. You can access this field once after creating. If you lost or forgot your api
...
    
...
    header:
    
...

...
  x-logo:
...
servers:
...
openapi: 3.0.0
paths:
...
  /v1/activity/list/suggestions/channels:
    get:
      operationId: ActivityController_getPublicChannels
      summary: List Suggested Channels
      description: List public channel suggestions
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
      responses:
        '200':
          description: User gets public channel suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicChannelSuggestionListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - Activity
  /v1/activity/list/suggestions/zone:
    get:
      operationId: ActivityController_getPublicZones
      summary: List Suggested Zones
      description: List public zone suggestions
      parameters:
        - name: limit
          required: false
          in: query
          description: The number of records to get.
          schema:
            default: 30
            type: number
        - name: skip
          required: false
          in: query
          description: The number of records to skip.
          schema:
            default: 0
            type: number
      responses:
        '200':
          description: User gets public zone suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicZoneSuggestionListResponse'
        '401':
          description: Error thrown when the user has not authorized or JWT is invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: You not authorized to use this route
                  error:
                    type: string
                    example: NOT_SIGNED_IN
      tags:
        - Activity
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PublicChannelSuggestionList:
      type: object
      properties:
        channel_id:
          type: number
        channel_createdOn:
          format: date-time
          type: string
        channel_name:
          type: string
        channel_description:
          type: string
        channel_public:
          type: boolean
        zone_id:
          type: number
        zone_name:
          type: string
        zone_subdomain:
          type: string
        channel_membersCount:
          type: string
      required:
        - channel_id
        - channel_createdOn
        - channel_name
        - channel_description
        - channel_public
        - zone_id
        - zone_name
        - zone_subdomain
        - channel_membersCount
    PublicChannelSuggestionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PublicChannelSuggestionList'
        total:
          type: number
        limit:
          type: number
        skip:
          type: number
      required:
        - data
        - total
        - limit
        - skip
    PublicZoneSuggestionList:
      type: object
      properties:
        zone_id:
          type: number
        zone_createdOn:
          format: date-time
          type: string
        zone_name:
          type: string
        zone_subdomain:
          type: string
        zone_description:
          type: string
        channel_public:
          type: boolean
        zone_channelCount:
          type: string
        zone_membersCount:
          type: string
      required:
        - zone_id
        - zone_createdOn
        - zone_name
        - zone_subdomain
        - zone_description
        - channel_public
        - zone_channelCount
        - zone_membersCount
    PublicZoneSuggestionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PublicZoneSuggestionList'
        total:
          type: number
        limit:
          type: number
        skip:
          type: number
      required:
        - data
        - total
        - limit
        - skip
...